fix: chris CRs - change h1 to h2 on plugin config page - remove `console.log` - move `AppInitializer`
