+++ src/components/AppInitializer.tsx
