+++ src/page/ConfigPageLayout/tabs/UninitializedTab.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { EmptyState, TextLink } from '@grafana/ui';
 
 import { ROUTES } from 'types';
-import { AppInitializer } from 'page/AppInitializer';
+import { AppInitializer } from 'components/AppInitializer';
 
 import { ConfigContent } from '../ConfigContent';
