+++ src/page/WelcomePage.tsx
@@ -5,9 +5,8 @@ import { Stack, Text, useStyles2 } from '@grafana/ui';
 import { css } from '@emotion/css';
 
 import { useMeta } from 'hooks/useMeta';
+import { AppInitializer } from 'components/AppInitializer';
 import { WelcomeTabs } from 'components/WelcomeTabs/WelcomeTabs';
-
-import { AppInitializer } from './AppInitializer';
 
 export const WelcomePage = () => {
   const styles = useStyles2(getStyles);
